pyproject.toml
[project]
name = "repro_mental_image_recon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "gdown>=5.2.0",
    "tqdm>=4.66.4",
    "mat73>=0.63",
    "omegaconf>=2.3.0",
    "einops>=0.8.0",
    "ftfy>=6.2.0",
    "pytorch-lightning>=2.3.1",
    "ipykernel>=6.29.5",
    "model-loads>=0.3",
    "transformers>=4.42.3",
    "scipy>=1.14.0",
    "matplotlib>=3.9.0",
    "numpy>=1.24.4",
    "matplotlib>=3.7.3",
    "seaborn>=0.13.0",
    "jupyter>=1.0.0",
    "ipympl>=0.9.3",
    "ipywidgets>=8.1.1",
    "bdpy @ git+https://github.com/KamitaniLab/bdpy@dev",
    "torchmetrics>=1.2.1",
    "pip>=24.1.2",
    "pyyaml>=6.0.1",
    "typing-extensions>=4.12.2",
    "timm==0.4.12",
    "dill>=0.3.8",
    "ninja>=1.11.1.1",
    "hydra-core>=1.3.2",
    "scikit-image>=0.24.0",
    "dreamsim>=0.2.1",
    "clip",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
